Окраска теперь в другом месте

+++ modules/global/src/com/company/marketplace/config/ColorConfig.java
@@ -6,13 +6,11 @@ import com.haulmont.cuba.core.config.Source;
 import com.haulmont.cuba.core.config.SourceType;
 import com.haulmont.cuba.core.config.defaults.Default;
 
-import java.math.BigDecimal;
-
 @Source(type = SourceType.DATABASE)
 public interface ColorConfig extends Config {
     @Property("color.price")
     @Default("100")
-    BigDecimal getProductPrice();
+    Integer getProductPrice();
 
-    void setProductPrice(BigDecimal price);
+    void setProductPrice(Integer price);
 }
